requirements.txt
Flask==3.0.0
joblib
numpy
pandas
PyYAML
requests
scikit-learn
boto3
